refactor: split CDP borrowing-revenue math helpers under the file-size cap
cdp-borrowing-revenue.ts crossed the 600-line soft cap (661). Extract the pure
math/format/bucket helpers (wei→token/USD conversion, day-bucket alignment,
window predicates, bucket accumulation) into cdp-borrowing-revenue-math.ts. The
public aggregate/series builders and exported types stay put, so no caller
import paths change. The math module only imports types from the main module
(type-only), so there is no circular runtime dependency.

Main file: 661 → 515 lines.
